[Mellanox] Improve the prompt of the Mellanox sdk sniffer command. (#546)
* [config/mlnx.py]
Prompt the risk of excessive consumption of disk space when user intends to turn on the sdk sniffer.
The SDK sniffer is a diagnosis feature, introduced with a view to recording the all the operations to SDK into the sniffer file which the low level team can replay and investigate. Only the sniffer file contains all the operations from SWSS restart does it help. To handle it in a way like log-rotate can hurt its integrity since the operations can be lost during rotating and to address this requires big effort.
So we just add a promot of that risk for now.

* [config/mlnx.py] adjust prompt.
adjust prompt in order to make it more clear.

* [config/mlnx.py] adjust prompt.
adjust prompt in order to make it more clear.
(recommit this in order to trigger vstest of the sonic-utilities and check whether it can pass)