[cli] Dynamic cli extension via plugins (#1186)
Added a mechanism to extend command line interface via plugins.
For every program - show, config, clear a new python package called 'plugins' is added. This package is used as a namespace for all plugins python modules. As an example mlnx.py is moved to the plugins package.

In the fututre, this mechanism will be used by Application Extension infrastructure.

Signed-off-by: Stepan Blyschak <stepanb@nvidia.com>