[fast-reboot] revert the change of disabling counter polling before fast-reboot (#1744)
What I did
The change being removed was introduced with PR#1174 with the condition that flex counter polling will be enabled in the boot up path unconditionally. However, a change was made to enable_counters.py to only enable counters when the configuration is missing (sonic-buildimage PR#7735). After the later change, the disabled flex counters will not be enabled when fast-reboot into the same image.

How to verify it
run fast-reboot test and snmp test, without the fix, the snmp test would fail. With the fix, the snmp test passes.

Signed-off-by: Ying Xie ying.xie@microsoft.com