[Show] Update the subcommands of Kdump. (#1682)
Signed-off-by: Yong Zhao yozhao@microsoft.com

What I did
This PR fixed the following issues:

Initially we need add prefix sudo when ran the show command of Kdump. This is unnecessary.

I re-organized the logic of implementing show command and also update the script sonic-kdump-config.

How I did it
Since the script sonic-kdump-config can only be executed by root user, I moved some functions from this file to
the kdump.py such that we do not need run the show command with sudo.

How to verify it
I verified this change on the DuT str-msn2700-03.