[fstrim] limit smartctl execution time to 30 seconds (#1850)
What I did
At early boot up time, smartctl is observed could stuck forever, which causes fstrim service to be stuck in pre-start state forever.
This is a transient issue. The periodic retry will go through when pmon docker is healthy.

How to verify it
This issue is not highly repeatable. Tested on the device hit the issue.

Signed-off-by: Ying Xie ying.xie@microsoft.com