[generic-config-updater] Cropping current/target config before sorting (#1831)
#### What I did
Fixing issue #1830

#### How I did it
Problem we consume the function `loadData` from sonic-yang-mgmt pkg that always crops tables without YANG models from configdb json object. It does it as a side-effect and is not an expected outcome of the function.

The fix here is to crop the current/target tables before doing any sorting, this way gurantee we avoid this bug. I think the better soln is to fix sonic-yang-mgmt pkg. Will look into this more next week.

#### How to verify it
Run the command in the issue, it will result in the expected outcome

#### Previous command output (if the output of a command-line utility has changed)

#### New command output (if the output of a command-line utility has changed)