Revert "[scripts/fast-reboot] Shutdown remaining containers through systemd (#2133)" (#2161)
This reverts commit 23e9398.

- What I did
Revert "[scripts/fast-reboot] Shutdown remaining containers through systemd (#2133)"

This reverted PR is part of a story that refactors warm/fast shutdown sequence to gracefully stop services instead of killing them without any ordering and dependency requirements which creates several issues and is error prone for the future.

This PR must come together with Azure/sonic-buildimage#10510.
However, #10510 is blocked due to an issue in swss-common Azure/sonic-swss-common#603
And a fix by MSFT is in review Azure/sonic-swss-common#606

I am reverting it because its dependency is still blocked and we cannot update submodule pointer. Once the dependency of the reverted PR is resolved, it shall be re-committed.