[VXLAN]Fixing traceback in show remotemac when mac moves during command execution (#2506)
- What I did
During the execution of show vxlan remotemac command, if a mac moves after the key is fetched but not the table fields, there will be the below traceback

show vxlan remotemac 2.2.2.2 
This is because the entry would have got removed before the CLI fetches the table.

- How I did it
Used get API to fetch table rather than indexing directly which resulted in traceback. Additionally fixed Python 3.8 SyntaxWarning.

show vxlan remotemac all
/usr/local/lib/python3.9/dist-packages/show/vxlan.py:101: SyntaxWarning: "is not" with a literal. Did you mean "!="?
if vtep_sip is not '0.0.0.0':
/usr/local/lib/python3.9/dist-packages/show/vxlan.py:108: SyntaxWarning: "is not" with a literal. Did you mean "!="?
if vtep_sip is not '0.0.0.0':

- How to verify it
Perform mac move during show command execution and verify there is no traceback.